+++ CHANGES.rst
@@ -4,9 +4,12 @@ Changelog
 5.0.10 (unreleased)
 -------------------
 
+- `search_sections` in IDesignPloneSettings has a new filed in each section:
+  `expandItems` that is a boolean to decide if the items of the section should 
+  be expanded or not (default is True).
+  [mamico]
 - Add a profile to limit addables on site root
   [lucabel]
-
 
 5.0.9 (2024-04-12)
 ------------------